The label in disassembly is generated from the address, not the length of a list array of existing labels. This has broken a lot of toys for some reason. Gotta figure out what's wrong there. I'll leave the TODO for now.
Removed commands outside the game memory area from disassembly.

Fixed formatting of disassembled code labels.

References to the memory area at the beginning of the program via `EQU` no longer refer to areas inside the program - it was causing a conflict because there are labels in the code for this purpose.   program - this was causing a conflict, since there are labels for this in `label:` code.